+++ tests/test_job_compiler_blender_render_progressive.py
@@ -130,13 +130,7 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
         mock_now = datetime.datetime.now(tz=tz_util.utc)
         mock_datetime.now.side_effect = [mock_now]
 
-        # We expect:
-        # - 1 destroy-intermediate task
-        # - 2 frame chunks x 3 sample chunks = 6 render tasks
-        # - 4 sample merge tasks
-        # - 1 'publish first chunk' task
-        # so a total of 12 tasks
-        task_ids = [ObjectId() for _ in range(50)]
+        task_ids = [ObjectId() for _ in range(19)]
         task_manager.api_create_task.side_effect = task_ids
 
         compiler = blender_render_progressive.BlenderRenderProgressive(
@@ -258,64 +252,24 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
             mock.call(  # task 7
                 job_doc,
                 [
-                    commands.MergeProgressiveRenders(
+                    commands.MergeProgressiveRenderSequence(
                         blender_cmd='/path/to/blender --enable-new-depsgraph',
                         input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-000001.exr',
                         input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-000001.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000001.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-######',
                         weight1=1,
                         weight2=9,
-                    ),
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-000002.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-000002.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000002.exr',
-                        weight1=1,
-                        weight2=9,
-                    ),
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-000003.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-000003.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000003.exr',
-                        weight1=1,
-                        weight2=9,
+                        frame_start=1,
+                        frame_end=5,
                     ),
                 ],
-                'merge-to-smpl10-frm1-3',
-                parents=[task_ids[1], task_ids[5]],
+                'merge-to-smpl10-frm1-5',
+                parents=[task_ids[1], task_ids[2], task_ids[5], task_ids[6]],
                 priority=-11,
                 status='under-construction',
                 task_type='exr-merge',
             ),
             mock.call(  # task 8
-                job_doc,
-                [
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-000004.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-000004.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000004.exr',
-                        weight1=1,
-                        weight2=9,
-                    ),
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/render-smpl-0001-0001-000005.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0002-0010-000005.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000005.exr',
-                        weight1=1,
-                        weight2=9,
-                    ),
-                ],
-                'merge-to-smpl10-frm4,5',
-                parents=[task_ids[2], task_ids[6]],
-                priority=-11,
-                status='under-construction',
-                task_type='exr-merge',
-            ),
-            mock.call(  # task 9
                 job_doc,
                 [commands.ExrSequenceToJpeg(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
@@ -325,11 +279,11 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 )],
                 'create-preview-images',
                 priority=-13,
-                parents=[task_ids[3], task_ids[7], task_ids[8]],
+                parents=[task_ids[3], task_ids[7]],
                 status='under-construction',
                 task_type='blender-render',
             ),
-            mock.call(  # task 10
+            mock.call(  # task 9
                 job_doc,
                 [commands.CreateVideo(
                     input_files='/render/out__intermediate-2018-07-06_115233/preview-*.jpg',
@@ -338,13 +292,13 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 )],
                 'create-preview-video',
                 priority=-13,
-                parents=[task_ids[4], task_ids[9]],
+                parents=[task_ids[4], task_ids[8]],
                 status='under-construction',
                 task_type='video-encoding',
             ),
 
             # Third Cycles chunk renders to intermediate directory.
-            mock.call(  # task 11
+            mock.call(  # task 10
                 job_doc,
                 [commands.BlenderRenderProgressive(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
@@ -361,7 +315,7 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 status='under-construction',
                 task_type='blender-render',
             ),
-            mock.call(  # task 12
+            mock.call(  # task 11
                 job_doc,
                 [commands.BlenderRenderProgressive(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
@@ -382,68 +336,28 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
             # Final merge pass. Could happen directly to the output directory, but to ensure the
             # intermediate directory shows a complete picture (pun intended), we take a similar
             # approach as earlier merge passes.
-            mock.call(  # task 13
+            mock.call(  # task 12
                 job_doc,
                 [
-                    commands.MergeProgressiveRenders(
+                    commands.MergeProgressiveRenderSequence(
                         blender_cmd='/path/to/blender --enable-new-depsgraph',
                         input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000001.exr',
                         input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-000001.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-000001.exr',
+                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-######',
                         weight1=10,
                         weight2=20,
-                    ),
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000002.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-000002.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-000002.exr',
-                        weight1=10,
-                        weight2=20,
-                    ),
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000003.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-000003.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-000003.exr',
-                        weight1=10,
-                        weight2=20,
+                        frame_start=1,
+                        frame_end=5,
                     ),
                 ],
-                'merge-to-smpl30-frm1-3',
-                parents=[task_ids[7], task_ids[11]],
-                priority=-21,
-                status='under-construction',
-                task_type='exr-merge',
-            ),
-            mock.call(  # task 14
-                job_doc,
-                [
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000004.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-000004.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-000004.exr',
-                        weight1=10,
-                        weight2=20,
-                    ),
-                    commands.MergeProgressiveRenders(
-                        blender_cmd='/path/to/blender --enable-new-depsgraph',
-                        input1='/render/out__intermediate-2018-07-06_115233/merge-smpl-0010-000005.exr',
-                        input2='/render/out__intermediate-2018-07-06_115233/render-smpl-0011-0030-000005.exr',
-                        output='/render/out__intermediate-2018-07-06_115233/merge-smpl-0030-000005.exr',
-                        weight1=10,
-                        weight2=20,
-                    ),
-                ],
-                'merge-to-smpl30-frm4,5',
-                parents=[task_ids[8], task_ids[12]],
+                'merge-to-smpl30-frm1-5',
+                parents=[task_ids[7], task_ids[10], task_ids[11]],
                 priority=-21,
                 status='under-construction',
                 task_type='exr-merge',
             ),
 
-            mock.call(  # task 15
+            mock.call(  # task 13
                 job_doc,
                 [commands.ExrSequenceToJpeg(
                     blender_cmd='/path/to/blender --enable-new-depsgraph',
@@ -453,11 +367,11 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 )],
                 'create-preview-images',
                 priority=-23,
-                parents=[task_ids[9], task_ids[13], task_ids[14]],
+                parents=[task_ids[8], task_ids[12]],
                 status='under-construction',
                 task_type='blender-render',
             ),
-            mock.call(  # task 16
+            mock.call(  # task 14
                 job_doc,
                 [commands.CreateVideo(
                     input_files='/render/out__intermediate-2018-07-06_115233/preview-*.jpg',
@@ -466,24 +380,24 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 )],
                 'create-preview-video',
                 priority=-23,
-                parents=[task_ids[10], task_ids[15]],
+                parents=[task_ids[9], task_ids[13]],
                 status='under-construction',
                 task_type='video-encoding',
             ),
 
-            mock.call(  # task 17
+            mock.call(  # task 15
                 job_doc,
                 [
                     commands.MoveOutOfWay(src='/render/out'),
                 ],
                 'move-outdir-out-of-way',
                 priority=-30,
-                parents=task_ids[13:15],
+                parents=[task_ids[12]],
                 status='under-construction',
                 task_type='file-management',
             ),
 
-            mock.call(  # task 18
+            mock.call(  # task 16
                 job_doc,
                 [
                     commands.CopyFile(
@@ -509,11 +423,11 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 ],
                 'publish-exr-to-output',
                 priority=-31,
-                parents=[task_ids[17]],
+                parents=[task_ids[15]],
                 status='under-construction',
                 task_type='file-management',
             ),
-            mock.call(  # task 19
+            mock.call(  # task 17
                 job_doc,
                 [
                     commands.CopyFile(
@@ -539,11 +453,11 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 ],
                 'publish-jpeg-to-output',
                 priority=-31,
-                parents=[task_ids[15], task_ids[17]],
+                parents=[task_ids[13], task_ids[15]],
                 status='under-construction',
                 task_type='file-management',
             ),
-            mock.call(  # task 20
+            mock.call(  # task 19
                 job_doc,
                 [
                     commands.CopyFile(
@@ -553,7 +467,7 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
                 ],
                 'publish-video-to-output',
                 priority=-31,
-                parents=[task_ids[16], task_ids[17]],
+                parents=[task_ids[14], task_ids[15]],
                 status='under-construction',
                 task_type='file-management',
             ),